CBV-88 Refactored BE code
