--- src/projectManager.py
@@ -1,26 +1,0 @@
-'''
-Last Updated 10/92/2022 by Montse
- 
-Will update project settings and state as more
-information becomes available to us
-
-Current state of project manager is barebones 
-skeleton containing definitions but no logic.
-Currently creates new project on createProject 
-API endpoint call
-'''
-import projectConfig
-
-class projectManager():
-
-    currentProject: projectConfig.project = None #Will create project object to store information off file
-    projectState = None #temp file call
-    
-    def __init__(self) -> None:
-        pass
-
-    def openProject():
-        return
-
-    def archiveProject():
-        return
