+++ src/routers/packetManager.py
@@ -1,5 +1,7 @@
 from dataSaver import dataSaver
+from dataGetter import dataGetter
 from fastapi import APIRouter
+from typing import Union
 
 router = APIRouter()
 
@@ -20,16 +22,6 @@ class packetManager():
 
 
     ##FUNCITONS
-
-    # this is hardcoded to read from the packets.txt file provided by the cutsomer for now, 
-    # until we can read the packets from the CAN Bus
-    def populatePacketList(self, projectId):
-        with open('packets.txt', 'r') as f:
-            for line in f:
-                fields = line.strip().split(';')
-                packet = {'projectId': projectId, 'timestamp': fields[0], 'type': fields[1], 'nodeId': fields[2], 'data': fields[3]}
-                self.packetList.append(packet)
-        dataSaver.storePackets(self.packetList)
 
     def savePacket(self, packet):
         return #ideally a status code to confirm it was saved, customer said memory will fill out fast
@@ -58,3 +50,19 @@ class packetManager():
     
     def generatePacket(self, listOfAttributes):
         ...
+
+    @router.get("/projects/{projectId}/packets", tags=["packets"])
+    def getPacketsFromProject(projectId: str, size: int, sort: str, node: Union[str, None] = None, before: Union[str, None] = None, after: Union[str, None] = None):
+        populatePacketList(projectId)
+        return dataGetter.getPackets(projectId, size, sort, node, before, after)
+
+# this is hardcoded to read from the packets.txt file provided by the cutsomer for now, 
+# until we can read the packets from the CAN Bus
+def populatePacketList(projectId):
+    packetList = []
+    with open('packets.txt', 'r') as f:
+        for line in f:
+            fields = line.strip().split(';')
+            packet = {'projectId': projectId, 'timestamp': fields[0], 'type': fields[1], 'nodeId': fields[2], 'data': fields[3]}
+            packetList.append(packet)
+    dataSaver.storePackets(packetList)
